+++ hydra-cardano-api/src/Hydra/Cardano/Api/TxIn.hs
@@ -1,11 +1,17 @@
+{-# OPTIONS_GHC -Wno-orphans #-}
+
 module Hydra.Cardano.Api.TxIn where
 
 import Hydra.Cardano.Api.Prelude
 
+import Cardano.Binary (unsafeDeserialize')
 import qualified Cardano.Ledger.Alonzo.TxInfo as Ledger
+import qualified Cardano.Ledger.BaseTypes as Ledger
 import qualified Cardano.Ledger.TxIn as Ledger
+import qualified Data.ByteString as BS
 import qualified Data.Set as Set
 import qualified PlutusLedgerApi.V2 as Plutus
+import Test.QuickCheck (choose, vectorOf)
 
 -- * Extras
 
@@ -49,3 +55,16 @@ fromPlutusTxOutRef (Plutus.TxOutRef (Plutus.TxId bytes) ix) =
 -- | Convert a cardano-api 'TxIn' into a plutus 'TxOutRef'.
 toPlutusTxOutRef :: TxIn -> Plutus.TxOutRef
 toPlutusTxOutRef = Ledger.txInfoIn' . toLedgerTxIn
+
+-- * Arbitrary values
+
+-- | A more random generator than the 'Arbitrary TxIn' from cardano-ledger.
+genTxIn :: Gen TxIn
+genTxIn =
+  fmap fromLedgerTxIn . Ledger.TxIn
+    -- NOTE: [88, 32] is a CBOR prefix for a bytestring of 32 bytes.
+    <$> fmap (unsafeDeserialize' . BS.pack . ([88, 32] <>)) (vectorOf 32 arbitrary)
+    <*> fmap Ledger.TxIx (choose (0, 99))
+
+instance Arbitrary TxIn where
+  arbitrary = genTxIn